+++ tcod/tcod.h
@@ -8,7 +8,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-int LineWhere(int x1, int y1, int x2, int y2, int *x_out, int *y_out);
+int bresenham(int x1, int y1, int x2, int y2, int n, int* __restrict out);
 #ifdef __cplusplus
 } // extern "C"
 #endif